if-def the event conversion code to fit with the ifdefs in WebCore. Review URL: http://codereview.chromium.org/1846
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@1908 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/glue/event_conversion.h
@@ -32,7 +32,7 @@ class MakePlatformMouseEvent : public WebCore::PlatformMouseEvent {
 
  private:
   static int last_click_count_;
-  static long last_click_time_;
+  static uint32 last_click_time_;
 };
 
 class MakePlatformWheelEvent : public WebCore::PlatformWheelEvent {
